Breaking Troops Testing in Seperate Py Files
